Transfer features about command execution codes to command package
