+++ src/main/scala/pro/civitaspo/digdag/plugin/ecs_task/command/EcsTaskCommandResultInternalOperator.scala
@@ -1,0 +1,12 @@
+package pro.civitaspo.digdag.plugin.ecs_task.command
+import io.digdag.client.config.Config
+import io.digdag.spi.{OperatorContext, TaskResult, TemplateEngine}
+import pro.civitaspo.digdag.plugin.ecs_task.AbstractEcsTaskOperator
+
+class EcsTaskCommandResultInternalOperator(operatorName: String, context: OperatorContext, systemConfig: Config, templateEngine: TemplateEngine)
+    extends AbstractEcsTaskOperator(operatorName, context, systemConfig, templateEngine) {
+
+  override def runTask(): TaskResult = {
+    null
+  }
+}
